+++ R/map_optimize.R
@@ -18,6 +18,8 @@
 #' @param titer_weights An optional matrix of weights to assign each titer when optimizing
 #' @param sort_optimizations Should optimizations be sorted by stress
 #'   afterwards?
+#' @param check_convergence Should a basic check for convergence of lowest stress
+#'   optimization runs onto a similar solution.
 #' @param verbose Should progress messages be reported, see also
 #'   `RacOptimizer.options()`
 #' @param options List of named optimizer options, see `RacOptimizer.options()`
@@ -53,6 +55,7 @@ optimizeMap <- function(
   ag_reactivity_adjustments = NULL,
   titer_weights = NULL,
   sort_optimizations = TRUE,
+  check_convergence = TRUE,
   verbose  = TRUE,
   options = list()
   ) {
@@ -121,6 +124,25 @@ optimizeMap <- function(
     "\n"
   )
 
+  # Check procrustes of the top 2 runs to see if there is much difference between them
+  if (check_convergence && numOptimizations(map) > 1) {
+
+    procrustes_dists <- c(
+      procrustesData(map, map, comparison_optimization_number = 2)$ag_dists,
+      procrustesData(map, map, comparison_optimization_number = 2)$sr_dists
+    )
+
+    if (max(procrustes_dists, na.rm = T) > 0.5) {
+      warning(sprintf(
+        singleline("There is some variation (%s AU for one point) in the top runs,
+                   this may be an indication that more optimization runs could help
+                   achieve a better optimum. Also consider running it with
+                   options = list(dim_annealing = TRUE).")
+      , round(max(procrustes_dists, na.rm = T), 2)))
+    }
+
+  }
+
   # Return the optimised map
   map
 
@@ -162,6 +184,7 @@ make.acmap <- function(
   minimum_column_basis    = "none",
   fixed_column_bases      = NULL,
   sort_optimizations      = TRUE,
+  check_convergence       = TRUE,
   verbose                 = TRUE,
   options                 = list(),
   ...
@@ -186,6 +209,7 @@ make.acmap <- function(
     minimum_column_basis = minimum_column_basis,
     fixed_column_bases = fixed_column_bases,
     sort_optimizations = sort_optimizations,
+    check_convergence = check_convergence,
     verbose = verbose,
     options = options
   )