+++ NEWS.md
@@ -75,3 +75,4 @@
 * Add option to `procrustesData()` to included only specific antigens or sera, as can already be done for `procrustesMap()`
 * Add methods to calculate point leverage, the effect of removing different strains and sera from the map, `agLeverage()`, `srLeverage()`, `titerLeverage()`
 * Change the `view.acmap()` method to allow for an option `num_optimizations` to specify how many optimizations to send to the viewer.
+* Add an option for and basic check by default for convergence of solutions when running map optimizations.